Loading kanji images fixed.
There was another problem with opening save files, it only saved the location of one image location to memory. This meant that whenever the user selected another kanji to load, it would stay on the first kanji (i.e., if I choose to load the kanji for "wind" and "student", it only loads the image for "wind" and even if I select "student", "wind" would be shown.

Now all kanji file locations are loaded.